+++ duels-plugin/src/main/java/me/realized/duels/arena/MatchImpl.java
@@ -28,7 +28,7 @@ public class MatchImpl implements Match {
     private final KitImpl kit;
     private final Map<UUID, List<ItemStack>> items;
     @Getter
-    private final int bet;
+    private final double bet;
     @Getter
     private final Queue source;
 
@@ -38,7 +38,7 @@ public class MatchImpl implements Match {
     // Default value for players is false, which is set to true if player is killed in the match.
     private final Map<Player, Boolean> players = new HashMap<>();
 
-    MatchImpl(final ArenaImpl arena, final KitImpl kit, final Map<UUID, List<ItemStack>> items, final int bet, final Queue source) {
+    MatchImpl(final ArenaImpl arena, final KitImpl kit, final Map<UUID, List<ItemStack>> items, final double bet, final Queue source) {
         this.arena = arena;
         this.start = System.currentTimeMillis();
         this.kit = kit;
